+++ edx_rest_framework_extensions/auth/jwt/tests/test_middleware.py
@@ -2,7 +2,7 @@
 Unit tests for jwt authentication middlewares.
 """
 from itertools import product
-from unittest.mock import ANY, Mock, patch
+from unittest.mock import Mock, patch
 
 import ddt
 from django.http.cookie import SimpleCookie
@@ -18,7 +18,6 @@ from rest_framework.views import APIView
 from rest_framework.viewsets import ViewSet
 from rest_framework_jwt.authentication import JSONWebTokenAuthentication
 
-from edx_rest_framework_extensions.auth.jwt.constants import USE_JWT_COOKIE_HEADER
 from edx_rest_framework_extensions.auth.jwt.cookies import (
     jwt_cookie_header_payload_name,
     jwt_cookie_name,
@@ -28,13 +27,8 @@ from edx_rest_framework_extensions.auth.jwt.middleware import (
     EnsureJWTAuthSettingsMiddleware,
     JwtAuthCookieMiddleware,
     JwtRedirectToLoginIfUnauthenticatedMiddleware,
-    _get_cached_user_from_jwt,
-    _get_user_from_jwt_original,
 )
-from edx_rest_framework_extensions.config import (
-    ENABLE_FORGIVING_JWT_COOKIES,
-    ENABLE_SET_REQUEST_USER_FOR_JWT_COOKIE,
-)
+from edx_rest_framework_extensions.config import ENABLE_SET_REQUEST_USER_FOR_JWT_COOKIE
 from edx_rest_framework_extensions.permissions import (
     IsStaff,
     IsSuperuser,
@@ -42,7 +36,6 @@ from edx_rest_framework_extensions.permissions import (
     LoginRedirectIfUnauthenticated,
     NotJwtRestrictedApplication,
 )
-from edx_rest_framework_extensions.settings import get_setting
 from edx_rest_framework_extensions.tests.factories import UserFactory
 
 
@@ -311,7 +304,7 @@ class TestJwtRedirectToLoginIfUnauthenticatedMiddleware(TestCase):
         ('/isauthenticatedandloginredirect/', False, 302),
         ('/isauthenticatedandloginredirect/', True, 200),
         ('/isauthenticated/', False, 401),
-        ('/isauthenticated/', True, 401),
+        ('/isauthenticated/', True, 200),
         ('/nopermissionsrequired/', False, 200),
         ('/nopermissionsrequired/', True, 200),
     )
@@ -328,8 +321,6 @@ class TestJwtRedirectToLoginIfUnauthenticatedMiddleware(TestCase):
     def test_login_required_middleware(self, url, has_jwt_cookies, expected_status):
         if has_jwt_cookies:
             self.client.cookies = _get_test_cookie()
-            if get_setting(ENABLE_FORGIVING_JWT_COOKIES):
-                expected_status = 200
         response = self.client.get(url)
         self.assertEqual(expected_status, response.status_code)
         if response.status_code == 302:
@@ -337,11 +328,11 @@ class TestJwtRedirectToLoginIfUnauthenticatedMiddleware(TestCase):
 
     @ddt.data(
         ('/loginredirectifunauthenticated/', False, 302),
-        ('/loginredirectifunauthenticated/', True, 302),
+        ('/loginredirectifunauthenticated/', True, 200),
         ('/isauthenticatedandloginredirect/', False, 302),
-        ('/isauthenticatedandloginredirect/', True, 302),
+        ('/isauthenticatedandloginredirect/', True, 200),
         ('/isauthenticated/', False, 401),
-        ('/isauthenticated/', True, 401),
+        ('/isauthenticated/', True, 200),
         ('/nopermissionsrequired/', False, 200),
         ('/nopermissionsrequired/', True, 200),
     )
@@ -358,20 +349,10 @@ class TestJwtRedirectToLoginIfUnauthenticatedMiddleware(TestCase):
     def test_login_required_overridden_middleware(self, url, has_jwt_cookies, expected_status):
         if has_jwt_cookies:
             self.client.cookies = _get_test_cookie()
-            if get_setting(ENABLE_FORGIVING_JWT_COOKIES):
-                expected_status = 200
         response = self.client.get(url)
         self.assertEqual(expected_status, response.status_code)
         if response.status_code == 302:
             self.assertEqual('/overridden/login/?next=' + url, response.url)
-
-
-# We want to duplicate these tests for now while we have two major code paths.  It will get unified once we have a
-# single way of doing JWT authentication again.
-@ddt.ddt
-@override_settings(EDX_DRF_EXTENSIONS={ENABLE_FORGIVING_JWT_COOKIES: True})
-class TestForgivingJwtRedirectToLoginIfUnauthenticatedMiddleware(TestJwtRedirectToLoginIfUnauthenticatedMiddleware):  # pylint: disable=test-inherits-tests # noqa E501 line too long
-    pass
 
 
 class CheckRequestUserForJwtAuthMiddleware(MiddlewareMixin):
@@ -402,12 +383,6 @@ class TestJwtAuthCookieMiddleware(TestCase):
         self.mock_response = Mock()
         self.middleware = JwtAuthCookieMiddleware(self.mock_response)
 
-    @patch('edx_rest_framework_extensions.auth.jwt.middleware.set_custom_attribute')
-    def test_do_not_use_jwt_cookies(self, mock_set_custom_attribute):
-        self.middleware.process_view(self.request, None, None, None)
-        self.assertIsNone(self.request.COOKIES.get(jwt_cookie_name()))
-        mock_set_custom_attribute.assert_any_call('use_jwt_cookie_requested', False)
-
     @ddt.data(
         (jwt_cookie_header_payload_name(), jwt_cookie_signature_name()),
         (jwt_cookie_signature_name(), jwt_cookie_header_payload_name()),
@@ -418,7 +393,6 @@ class TestJwtAuthCookieMiddleware(TestCase):
     def test_missing_cookies(
             self, set_cookie_name, missing_cookie_name, mock_set_custom_attribute, mock_log
     ):
-        self.request.META[USE_JWT_COOKIE_HEADER] = 'true'
         self.request.COOKIES[set_cookie_name] = 'test'
         self.middleware.process_view(self.request, None, None, None)
         self.assertIsNone(self.request.COOKIES.get(jwt_cookie_name()))
@@ -426,99 +400,25 @@ class TestJwtAuthCookieMiddleware(TestCase):
             '%s cookie is missing. JWT auth cookies will not be reconstituted.' %
             missing_cookie_name
         )
-        mock_set_custom_attribute.assert_any_call('use_jwt_cookie_requested', True)
         mock_set_custom_attribute.assert_any_call('has_jwt_cookie', False)
 
     @patch('edx_rest_framework_extensions.auth.jwt.middleware.set_custom_attribute')
     def test_no_cookies(self, mock_set_custom_attribute):
-        self.request.META[USE_JWT_COOKIE_HEADER] = 'true'
         self.middleware.process_view(self.request, None, None, None)
         self.assertIsNone(self.request.COOKIES.get(jwt_cookie_name()))
-        mock_set_custom_attribute.assert_any_call('use_jwt_cookie_requested', True)
         mock_set_custom_attribute.assert_any_call('has_jwt_cookie', False)
 
     @patch('edx_rest_framework_extensions.auth.jwt.middleware.set_custom_attribute')
     def test_success(self, mock_set_custom_attribute):
-        self.request.META[USE_JWT_COOKIE_HEADER] = 'true'
         self.request.COOKIES[jwt_cookie_header_payload_name()] = 'header.payload'
         self.request.COOKIES[jwt_cookie_signature_name()] = 'signature'
         self.middleware.process_view(self.request, None, None, None)
         self.assertEqual(self.request.COOKIES[jwt_cookie_name()], 'header.payload.signature')
-        mock_set_custom_attribute.assert_any_call('use_jwt_cookie_requested', True)
         mock_set_custom_attribute.assert_any_call('has_jwt_cookie', True)
 
     _LOG_WARN_AUTHENTICATION_FAILED = 0
     _LOG_WARN_MISSING_JWT_AUTHENTICATION_CLASS = 1
 
-    @patch('edx_rest_framework_extensions.auth.jwt.middleware.log')
-    @patch(
-        'edx_rest_framework_extensions.auth.jwt.middleware._get_cached_user_from_jwt',
-        wraps=_get_cached_user_from_jwt
-    )
-    @patch(
-        'edx_rest_framework_extensions.auth.jwt.middleware._get_user_from_jwt_original',
-        wraps=_get_user_from_jwt_original
-    )
-    @ddt.data(
-        ('/nopermissionsrequired/', True, True, True, None, False),
-        ('/nopermissionsrequired/', True, False, False, None, False),
-        ('/nopermissionsrequired/', False, False, True, _LOG_WARN_AUTHENTICATION_FAILED, False),
-        ('/nopermissionsrequired/', False, False, False, None, False),
-        ('/nopermissionsrequired/', True, True, True, None, True),
-        ('/unauthenticated/', True, False, True, _LOG_WARN_MISSING_JWT_AUTHENTICATION_CLASS, False),
-        ('/unauthenticated/', True, False, True, None, False),
-    )
-    @ddt.unpack
-    def test_set_request_user_with_use_jwt_cookie(
-            self, url, is_cookie_valid, is_request_user_set, is_toggle_enabled, log_warning, send_post,
-            mock_protected_get_user_original, mock_protected_get_cached_user_from_jwt, mock_log,
-    ):
-        header = {USE_JWT_COOKIE_HEADER: 'true'}
-        self.client.cookies = _get_test_cookie(is_cookie_valid=is_cookie_valid)
-        check_user_middleware_assertion_class = (
-            'CheckRequestUserForJwtAuthMiddleware'
-            if is_request_user_set else
-            'CheckRequestUserAnonymousForJwtAuthMiddleware'
-        )
-        with override_settings(
-            ROOT_URLCONF='edx_rest_framework_extensions.auth.jwt.tests.test_middleware',
-            MIDDLEWARE=(
-                    'django.contrib.sessions.middleware.SessionMiddleware',
-                    'edx_rest_framework_extensions.auth.jwt.middleware.JwtAuthCookieMiddleware',
-                    'django.contrib.auth.middleware.AuthenticationMiddleware',
-                    'edx_rest_framework_extensions.auth.jwt.tests.test_middleware.{}'.format(
-                        check_user_middleware_assertion_class
-                    ),
-            ),
-            EDX_DRF_EXTENSIONS={
-                ENABLE_FORGIVING_JWT_COOKIES: False,
-                ENABLE_SET_REQUEST_USER_FOR_JWT_COOKIE: is_toggle_enabled,
-            }
-        ):
-            if send_post:
-                response = self.client.post(url, {}, content_type="application/json", **header)
-            else:
-                response = self.client.get(url, **header)
-            self.assertEqual(200, response.status_code)
-
-        if log_warning == self._LOG_WARN_AUTHENTICATION_FAILED:
-            mock_log.debug.assert_called_once_with('Jwt Authentication failed and request.user could not be set.')
-        elif log_warning == self._LOG_WARN_MISSING_JWT_AUTHENTICATION_CLASS:
-            mock_log.debug.assert_called_once_with(
-                'Jwt Authentication expected, but view %s is not using a JwtAuthentication class.', ANY
-            )
-        else:
-            mock_log.warn.assert_not_called()
-        # Ensure _get_user_from_jwt_original is called, since ENABLE_FORGIVING_JWT_COOKIES is disabled
-        if is_request_user_set:
-            mock_protected_get_user_original.assert_called()
-        # Ensure _get_cached_user_from_jwt is not called, since ENABLE_FORGIVING_JWT_COOKIES is disabled
-        mock_protected_get_cached_user_from_jwt.assert_not_called()
-
-    @patch(
-        'edx_rest_framework_extensions.auth.jwt.middleware._get_cached_user_from_jwt',
-        wraps=_get_cached_user_from_jwt
-    )
     @override_settings(
         ROOT_URLCONF='edx_rest_framework_extensions.auth.jwt.tests.test_middleware',
         MIDDLEWARE=(
@@ -527,16 +427,11 @@ class TestJwtAuthCookieMiddleware(TestCase):
                 'django.contrib.auth.middleware.AuthenticationMiddleware',
                 'edx_rest_framework_extensions.auth.jwt.tests.test_middleware.CheckRequestUserForJwtAuthMiddleware',
         ),
-        EDX_DRF_EXTENSIONS={
-            ENABLE_FORGIVING_JWT_COOKIES: True,
-            ENABLE_SET_REQUEST_USER_FOR_JWT_COOKIE: True,
-        }
+        EDX_DRF_EXTENSIONS={ENABLE_SET_REQUEST_USER_FOR_JWT_COOKIE: True}
     )
-    def test_set_request_user_with_use_jwt_cookie_and_forgiving_jwt_cookie_enabled(
-        self, mock_protected_get_cached_user_from_jwt
-    ):
+    def test_set_request_user_with_use_jwt_cookie_middleware(self):
         """
-        Tests set request user when ENABLE_FORGIVING_JWT_COOKIES is also enabled.
+        Tests set request user with the JwtAuthCookieMiddleware.
         """
         self.client.cookies = _get_test_cookie(is_cookie_valid=True)
 
@@ -544,16 +439,6 @@ class TestJwtAuthCookieMiddleware(TestCase):
 
         # additional assertions in CheckRequestUserForJwtAuthMiddleware
         self.assertEqual(200, response.status_code)
-        # Ensure _get_cached_user_from_jwt is called, since ENABLE_FORGIVING_JWT_COOKIES is enabled
-        mock_protected_get_cached_user_from_jwt.assert_called()
-
-
-# We want to duplicate these tests for now while we have two major code paths.  It will get unified once we have a
-# single way of doing JWT authentication again.
-@ddt.ddt
-@override_settings(EDX_DRF_EXTENSIONS={ENABLE_FORGIVING_JWT_COOKIES: True})
-class TestForgivingJwtAuthCookieMiddleware(TestJwtAuthCookieMiddleware):  # pylint: disable=test-inherits-tests
-    pass
 
 
 def _get_test_cookie(is_cookie_valid=True):