+++ edx_rest_framework_extensions/tests/test_middleware.py
@@ -9,7 +9,6 @@ from django.contrib.auth.models import AnonymousUser
 from django.test import RequestFactory, TestCase
 from edx_django_utils.cache import RequestCache
 
-from edx_rest_framework_extensions.auth.jwt.constants import USE_JWT_COOKIE_HEADER
 from edx_rest_framework_extensions.auth.jwt.cookies import jwt_cookie_name
 from edx_rest_framework_extensions.middleware import (
     RequestCustomAttributesMiddleware,
@@ -112,7 +111,6 @@ class TestRequestCustomAttributesMiddleware(TestCase):
     @patch('edx_django_utils.monitoring.set_custom_attribute')
     def test_request_auth_type_guess_jwt_cookie_attribute(self, mock_set_custom_attribute):
         self.request.user = UserFactory()
-        self.request.META[USE_JWT_COOKIE_HEADER] = True
         self.request.COOKIES[jwt_cookie_name()] = 'reconstituted-jwt-cookie'
 
         self.middleware.process_response(self.request, None)